fix: add a dialog for delete and edit
Changed to use a new widget for delete and edit of the note.

The previous approach was to use the context of `NoteSheet` for showing
indicators or routing.
The approach does not work well for some cases because the operation
deletes the note `NoteSheet` is watching for, and the context is
unmounted before pushing to the post page.
With this PR, adding `DeleteAndEditDialog`, a separate context is used
during delete and edit so that the context will not be unmounted.