feat: add SwRegisterDialog
The `sw/register` endpoint of Misskey is marked `secure` and cannot be
accessed using an access token issued with MiAuth.
This dialog generates an AiScript code that sends the request and
prompts the user to run it on the scratchpad of Misskey Web.
Once the code is executed, the scratchpad shows a dialog with the
response.
Finally, by pasting the response, Aria can indirectly utilise the
endpoint.