feat: subscribe to push notification
Added a page to register to push notifications.
When the switch is enabled, it generates keys that are used to later
decrypt Web Push messages sent from Misskey.
The keys and an endpoint to which Misskey server sends notifications are
then sent.

The endpoint must be unique among all subscriptions.
If a user is on an Android device and chooses an external UnifiedPush
distributor like ntfy, that distributor will provide the endpoint.
In this case, encrypted Web Push messages will be distributed, and Aria
needs to decrypt that within the app, so the keys will be saved on the
device.
If the user doesn't have any distributor apps, chooses the internal
distributor, or is on iOS, the endpoint directs to [Misskey Web Push
Proxy](https://github.com/poppingmoon/misskey-web-push-proxy).
The keys will be sent to the proxy to allow it to decrypt the messages
on the server.
Generation of the endpoint will be implemented in later commits.

After that, the endpoint will be stored on the device for later
unsubscription.