feat: use EmbeddedDistributor to get FCM token
Added [Android FOSS Embedded FCM Distributor]
(https://codeberg.org/UnifiedPush/android-embedded_fcm_distributor) for
a fallback of UnifiedPush Distributors.
If the user does not have any distributor apps installed, this will be
automatically selected.
Otherwise, it will appear as one of a distributors that the user can
choose.

According to the official documentation, the `getEndpoint` function is
supposed to return an endpoint.
However, for Aria, it has been modified to return an FCM token instead,
since the actual endpoint will be generated on the Flutter side.

The token is prefixed to indicate that the returned value is from the
embedded distributor or from an external distributor.
On the Flutter side, the prefix is removed and the token will be sent to
the proxy to allow it to send notifications via FCM.