feat: save user id of the account
Added the `UserIdsNotifier` which saves the `userId` associated with the
account.
This will be used to determine which account the notification was sent
to.

The function to save the id will be called on signing in and subscribing
to push notifications.
This ensures that the id is available when a push notification is
received, even if the user was signed in before this feature was added.